git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/ERM.BL/Dev/Features/ActivityMigration/BL;C24247

+++ Source/ApplicationServices/Reports/2Gis.Erm.BL.Reports/PlanningReport/Лист_ДЗ_накопленная_с_01_12.sql
@@ -1,5 +1,5 @@
 ﻿--declare 	@IssueDate date = '20130901'
---			,@City int = 1
+--			,@City bigint = 1
 DECLARE
 	@DZCheckPoint date = '20111201'
 
@@ -7,7 +7,7 @@ DECLARE
 DECLARE 
 	@StartDate DATE = @DZCheckPoint
 	, @EndDate DATE = @IssueDate
-	, @OrganizationUnit INT = @City
+	, @OrganizationUnit bigint = @City
 	, @ShowLimits BIT = 1;
 
 --Отсюда начинается запрос отчета по оборотам лицевых счетов
